use full uuid as a string. convert all indices to strings

+++ ecoli/processes/transcript_initiation.py
@@ -438,7 +438,7 @@ class TranscriptInitiation(PartitionedProcess):
 
         # new RNAPs
         RNAP_indexes = create_unique_indexes(n_RNAPs_to_activate)
-        RNAP_indexes = np.array(RNAP_indexes).astype(int)
+        RNAP_indexes = np.array(RNAP_indexes)
         new_RNAPs = arrays_to(
             n_RNAPs_to_activate, {
                 'unique_index': RNAP_indexes,
@@ -455,7 +455,7 @@ class TranscriptInitiation(PartitionedProcess):
         # Add partially transcribed RNAs
         is_mRNA = np.isin(TU_index_partial_RNAs, self.idx_mRNA)
         rna_indices = create_unique_indexes(n_RNAPs_to_activate)
-        rna_indices = np.array(rna_indices).astype(int)
+        rna_indices = np.array(rna_indices)
         new_RNAs = arrays_to(
             n_RNAPs_to_activate, {
                 'unique_index': rna_indices,
